[front] - refactor: reorganize MCP components and forms
 - Moved MCP forms related components to a dedicated forms directory
 - Adjusted import paths to align with the new directory structure for MCP components
