+++ front/components/actions/mcp/create/CustomHeadersConfigurationSection.tsx
@@ -12,8 +12,8 @@ import {
   useWatch,
 } from "react-hook-form";
 
+import type { CreateMCPServerDialogFormValues } from "@app/components/actions/mcp/forms/types";
 import { McpServerHeaders } from "@app/components/actions/mcp/MCPServerHeaders";
-import type { CreateMCPServerDialogFormValues } from "@app/components/actions/mcp/types";
 import { requiresBearerTokenConfiguration } from "@app/lib/actions/mcp_helper";
 import type { DefaultRemoteMCPServerConfig } from "@app/lib/actions/mcp_internal_actions/remote_servers";
 import type { MCPServerType } from "@app/lib/api/mcp";
