+++ front/components/actions/mcp/create/MCPServerViewForm.tsx
