+++ front/components/actions/mcp/create/RemoteMCPServerConfigurationSection.tsx
@@ -13,7 +13,7 @@ import {
 } from "@dust-tt/sparkle";
 import { useController, useFormContext } from "react-hook-form";
 
-import type { CreateMCPServerDialogFormValues } from "@app/components/actions/mcp/types";
+import type { CreateMCPServerDialogFormValues } from "@app/components/actions/mcp/forms/types";
 import type { DefaultRemoteMCPServerConfig } from "@app/lib/actions/mcp_internal_actions/remote_servers";
 import type { AuthorizationInfo } from "@app/lib/actions/mcp_metadata_extraction";
