+++ front/components/actions/mcp/forms/submitConnectMCPServerDialogForm.ts
@@ -1,4 +1,4 @@
-import type { MCPServerOAuthFormValues } from "@app/components/actions/mcp/types";
+import type { MCPServerOAuthFormValues } from "@app/components/actions/mcp/forms/types";
 import { getMcpServerDisplayName } from "@app/lib/actions/mcp_helper";
 import type { AuthorizationInfo } from "@app/lib/actions/mcp_metadata_extraction";
 import type { MCPServerViewType } from "@app/lib/api/mcp";
