+++ front/components/actions/mcp/forms/submitCreateMCPServerDialogForm.ts
@@ -1,4 +1,4 @@
-import type { CreateMCPServerDialogFormValues } from "@app/components/actions/mcp/types";
+import type { CreateMCPServerDialogFormValues } from "@app/components/actions/mcp/forms/types";
 import { requiresBearerTokenConfiguration } from "@app/lib/actions/mcp_helper";
 import type { AuthorizationInfo } from "@app/lib/actions/mcp_metadata_extraction";
 import type { MCPServerType } from "@app/lib/api/mcp";
